+++ bin-proto/src/discriminable.rs
@@ -1,3 +1,4 @@
+/// A trait for types with discriminants (typically Enums).
 pub trait Discriminable {
     type Discriminant;