+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # v0.5.0
 - Split `Protocol` into `ProtocolRead` and `ProtocolWrite`
-- Convert `ExternallyLengthPrefixed` to `ExternallyTaggedRead`
+- Split `ExternallyLengthPrefixed` into `ExternallyTaggedRead` and `ExternallyTaggedWrite`
 - Convert `FlexibleArrayMember` to `FlexibleArrayMemberRead`
 - Split `BitField` into `BitFieldWrite` and `BitFieldRead`
-- Implement `ExternallyTaggedRead` and `FlexibleArrayMemberRead` on all list and map types and `Option`
+- Implement `ExternallyTaggedRead`, `ExternallyTaggedWrite` `FlexibleArrayMemberRead` on all list and map types and `Option`
 - Add `Error` variant for failed `TryFrom` conversion for `ExternallyLengthPrefixed` tags
 - Add generic `Tag` parameter to `ExternallyTaggedRead`
 - Allow for `#[protocol(tag(type = "<type>", write_value = "<expr>"))]` attribute macro